AdaptiveGAN
This is a Generative Adversarial Networks implementation with the adaptive distribution of generative and discriminative network models. The models are trained on an adaptive neural network with varying hidden layers to  visualise the training of discriminative and generative models to generate a gaussian distribution with mean 0 and variance 1 more effectively.